Implement help menu

+++ calculator.h
@@ -1,6 +1,6 @@
 int add(int, int);
 int multiply(int, int);
-int divide(int, int);
+float divide(int, int);
 int subtract(int, int);
 int modulus(int, int);
 int main(int, char**);
